+++ errors/miss_factory_error.py
@@ -3,5 +3,8 @@ from .base_bot_error	import BaseBotError
 
 class MissFactoryError(BaseBotError):
 	
-	"""我是doc(不知道写啥"""
+	"""
+	ensure_session 未命中缓存 且调用方没有传入建连工厂
+	此时无法凭空建出 PiClient 只能报错
+	"""
 	...